pc_com/requirements.txt
protobuf==3.20.1
pyserial==3.5
PySide6_Essentials==6.7.2
shiboken6==6.7.2
numpy==2.2.1
pyqtgraph==0.13.7